+++ src/components/Button/Button.tsx
@@ -27,7 +27,6 @@ const ButtonContainer = styled.a<ButtonProps>`
 
           &:active {
             background-color: ${theme.colors.accent_primary_0};
-            border-color: ${theme.colors.accent_primary_0};
           }
         `;
       case ButtonStyles.secondary:
@@ -38,7 +37,6 @@ const ButtonContainer = styled.a<ButtonProps>`
           box-shadow: ${theme.shadows.default};
           &:active {
             background-color: ${theme.colors.accent_secondary_0};
-            border-color: ${theme.colors.accent_secondary_0};
           }
         `;
       case ButtonStyles.success:
@@ -49,7 +47,6 @@ const ButtonContainer = styled.a<ButtonProps>`
           box-shadow: ${theme.shadows.default};
           &:active {
             background-color: ${theme.colors.accent_success_0};
-            border-color: ${theme.colors.accent_success_0};
           }
         `;
       case ButtonStyles.info:
@@ -60,7 +57,6 @@ const ButtonContainer = styled.a<ButtonProps>`
           box-shadow: ${theme.shadows.default};
           &:active {
             background-color: ${theme.colors.accent_info_0};
-            border-color: ${theme.colors.accent_info_0};
           }
         `;
       case ButtonStyles.warning:
@@ -71,7 +67,6 @@ const ButtonContainer = styled.a<ButtonProps>`
           box-shadow: ${theme.shadows.default};
           &:active {
             background-color: ${theme.colors.accent_warning_0};
-            border-color: ${theme.colors.accent_warning_0};
           }
         `;
       case ButtonStyles.danger:
@@ -82,12 +77,12 @@ const ButtonContainer = styled.a<ButtonProps>`
           box-shadow: ${theme.shadows.default};
           &:active {
             background-color: ${theme.colors.accent_danger_0};
-            border-color: ${theme.colors.accent_danger_0};
           }
         `;
       case ButtonStyles.transparent:
         return `
           color: ${theme.colors.text.dark};
+          background-color: transparent;
 
           &:active {
             background-color: ${theme.colors.transparency.black(0.1)};
@@ -125,10 +120,11 @@ const ButtonContainer = styled.a<ButtonProps>`
     transition-duration: ${theme.transition.times.s};
   `}
   border-radius: 100vmax;
+  border: none;
 
   font-weight: 500;
 
-  transition-property: background-color, border-color, color;
+  transition-property: background-color, color;
   transition-timing-function: ease-in-out;
   text-decoration: none;
   justify-content: center;
@@ -139,17 +135,30 @@ const ButtonContainer = styled.a<ButtonProps>`
 /**
  * Button and Icon Button Component
  */
-const Button = ({children, className, href=undefined, isIconButton=false, style=ButtonStyles.default, padding, gap, width, title, onClick}:ButtonType) => {
+const Button = ({
+  children, 
+  href=undefined,
+  as='a',
+  type='button',
+  isIconButton=false, 
+  style=ButtonStyles.default, 
+  padding, 
+  gap, 
+  width, 
+  title, 
+  onClick,
+  ...rest
+}:ButtonType) => {
   const propagateClick = (event: React.MouseEvent) => {
-    event.preventDefault();
     if(typeof navigator.vibrate === 'function') navigator.vibrate(20);
     if(onClick) onClick(event);
   };
   return (
     <ButtonContainer 
       children={children}
+      as={as}
       href={href} 
-      className={className}
+      type={type}
       $style={style}
       $padding={padding}
       $gap={gap}
@@ -157,6 +166,7 @@ const Button = ({children, className, href=undefined, isIconButton=false, style=
       $isIconButton={isIconButton}
       onClick={propagateClick}
       title={title}
+      {...rest}
     />
   );
 };