+++ src/components/ProfileButton/ProfileButton.type.ts
@@ -6,7 +6,7 @@
  * <ProfileButton src='profile.jpg' alt='Profile' onClick={handleClick} />
  */
 export type ProfileButtonType = {
-  src: string;
+  src?: string;
   alt?: string;
   size?: string;
   onClick?: (event?:React.MouseEvent) => void;