+++ src/components/QuoteDialog/QuoteDialog.type.ts
@@ -1,0 +1,8 @@
+import { DialogType } from "components/Dialog/Dialog.type";
+
+/**
+ * QuoteDialogType type for QuoteDialog component
+ * @example
+ * <QuoteDialog open={true} toggleDialog={() => {}} />
+ */
+export type QuoteDialogType = Omit<DialogType, 'children'> & {};